+++ RELEASE-NOTES.md
@@ -22,6 +22,8 @@
 ## New additions
 * Added connection option `--token-file-path` allowing passing OAuth token using a file. The function is also
   supported by setting `token_file_path` in connection definition.
+* Support for Python remote execution via `snow stage execute` and `snow git execute` similar to existing EXECUTE IMMEDIATE support.
+* Added support for autocomplete in `--connection` flag.
 
 ## Fixes and improvements
 * The `snow app run` command now allows upgrading to unversioned mode from a versioned or release mode application installation
@@ -29,7 +31,18 @@
 * The `snow app version create` command now allows operating on application packages created outside the CLI
 * Added support for user stages in stage execute command
 * Added support for user stages in stage and git copy commands
+* Improved support for quoted identifiers in snowpark commands.
 
+# v2.6.1
+## Backward incompatibility
+
+## Deprecations
+
+## New additions
+
+## Fixes and improvements
+* `snow object create` message returns meaningful error if connection database is not defined.
+* Fixed crashing when save_logs is false and log directory does not exist
 
 # v2.6.0
 ## Backward incompatibility